Don't crash on ENOENT when adding watch inside event_gen
If we get an event notification about a directory being moved and then
it disappears before we have time to add a watch to it, we need to
catch the resulting ENOENT error and warn about it rather than
crashing.

This is similar to the fix made in
ef6769dda4c7f1396cea5f8c5a5530b23b81558f.