InotifyTree/InotifyTrees should not follow symlinks by default
It is almost certainly not the intention of the user to follow
symbolic links to directories when using InotifyTree / InotifyTrees,
so default to not doing that.

*** This is a change in behavior. *** I believe it is the correct
thing to do, but if it would be preferred to provide the ability not
to follow symlinks without changing the behavior by default, then the
way to do that is to change `follow_symlinks=False` to
`follow_symlinks=True` in the `__init__` function definitions for
`_BaseTree`, `InotifyTree`, and `InotifyTrees`.