Fix several bugs in handling directory move events
1) When a directory is moved from inside a directory we watch, to a
different directory that we don't watch, then we get a MOVED_FROM
event but no MOVED_TO event, in which case we need to clean up the
watch for that directory because we are no longer supposed to be
watching it.

2) When a directory is moved from inside a directory we watch to inside
another directory we watch, then we need to update our internal data
to reflect the new location of the directory.

3) For both of the above, we need to handle subdirectories of moved
directories, not just the directories themselves.

To facilitate these fixes, we:

* Add a new `watches()` iterator to the `Inotify` class so the tree
  classes can iterate through all watches to find ones that need to be
  removed.
* Add a new `remove_tree()` function to the tree classes which handles
  removing entire trees.

Both of these new functions are available to end users of the library
since they're useful and there's no reason for them not to be.

See the big comment added to the code for additional details about how
this fix is implemented.